Conflicts:
	doc/03-cli.md
	doc/04-schema.md
	src/Composer/Factory.php

+++ src/Composer/Installer/LibraryInstaller.php
@@ -14,10 +14,10 @@ namespace Composer\Installer;
 
 use Composer\Composer;
 use Composer\IO\IOInterface;
-use Composer\Downloader\DownloadManager;
 use Composer\Repository\InstalledRepositoryInterface;
 use Composer\Package\PackageInterface;
 use Composer\Util\Filesystem;
+use Composer\Util\ProcessExecutor;
 
 /**
  * Package installation manager.
@@ -41,15 +41,16 @@ class LibraryInstaller implements InstallerInterface
      * @param IOInterface $io
      * @param Composer    $composer
      * @param string      $type
+     * @param Filesystem  $filesystem
      */
-    public function __construct(IOInterface $io, Composer $composer, $type = 'library')
+    public function __construct(IOInterface $io, Composer $composer, $type = 'library', Filesystem $filesystem = null)
     {
         $this->composer = $composer;
         $this->downloadManager = $composer->getDownloadManager();
         $this->io = $io;
         $this->type = $type;
 
-        $this->filesystem = new Filesystem();
+        $this->filesystem = $filesystem ?: new Filesystem();
         $this->vendorDir = rtrim($composer->getConfig()->get('vendor-dir'), '/');
         $this->binDir = rtrim($composer->getConfig()->get('bin-dir'), '/');
     }
@@ -116,20 +117,17 @@ class LibraryInstaller implements InstallerInterface
     public function uninstall(InstalledRepositoryInterface $repo, PackageInterface $package)
     {
         if (!$repo->hasPackage($package)) {
-            // TODO throw exception again here, when update is fixed and we don't have to remove+install (see #125)
-            return;
             throw new \InvalidArgumentException('Package is not installed: '.$package);
         }
-
-        $downloadPath = $this->getInstallPath($package);
 
         $this->removeCode($package);
         $this->removeBinaries($package);
         $repo->removePackage($package);
 
+        $downloadPath = $this->getPackageBasePath($package);
         if (strpos($package->getName(), '/')) {
             $packageVendorDir = dirname($downloadPath);
-            if (is_dir($packageVendorDir) && !glob($packageVendorDir.'/*')) {
+            if (is_dir($packageVendorDir) && $this->filesystem->isDirEmpty($packageVendorDir)) {
                 @rmdir($packageVendorDir);
             }
         }
@@ -140,10 +138,16 @@ class LibraryInstaller implements InstallerInterface
      */
     public function getInstallPath(PackageInterface $package)
     {
-        $this->initializeVendorDir();
         $targetDir = $package->getTargetDir();
 
-        return ($this->vendorDir ? $this->vendorDir.'/' : '') . $package->getPrettyName() . ($targetDir ? '/'.$targetDir : '');
+        return $this->getPackageBasePath($package) . ($targetDir ? '/'.$targetDir : '');
+    }
+
+    protected function getPackageBasePath(PackageInterface $package)
+    {
+        $this->initializeVendorDir();
+
+        return ($this->vendorDir ? $this->vendorDir.'/' : '') . $package->getPrettyName();
     }
 
     protected function installCode(PackageInterface $package)
@@ -154,13 +158,28 @@ class LibraryInstaller implements InstallerInterface
 
     protected function updateCode(PackageInterface $initial, PackageInterface $target)
     {
-        $downloadPath = $this->getInstallPath($initial);
-        $this->downloadManager->update($initial, $target, $downloadPath);
+        $initialDownloadPath = $this->getInstallPath($initial);
+        $targetDownloadPath = $this->getInstallPath($target);
+        if ($targetDownloadPath !== $initialDownloadPath) {
+            // if the target and initial dirs intersect, we force a remove + install
+            // to avoid the rename wiping the target dir as part of the initial dir cleanup
+            if (substr($initialDownloadPath, 0, strlen($targetDownloadPath)) === $targetDownloadPath
+                || substr($targetDownloadPath, 0, strlen($initialDownloadPath)) === $initialDownloadPath
+            ) {
+                $this->removeCode($initial);
+                $this->installCode($target);
+
+                return;
+            }
+
+            $this->filesystem->rename($initialDownloadPath, $targetDownloadPath);
+        }
+        $this->downloadManager->update($initial, $target, $targetDownloadPath);
     }
 
     protected function removeCode(PackageInterface $package)
     {
-        $downloadPath = $this->getInstallPath($package);
+        $downloadPath = $this->getPackageBasePath($package);
         $this->downloadManager->remove($package, $downloadPath);
     }
 
@@ -178,9 +197,15 @@ class LibraryInstaller implements InstallerInterface
         foreach ($binaries as $bin) {
             $binPath = $this->getInstallPath($package).'/'.$bin;
             if (!file_exists($binPath)) {
-                $this->io->write('    <warning>Skipped installation of '.$bin.' for package '.$package->getName().': file not found in package</warning>');
+                $this->io->write('    <warning>Skipped installation of bin '.$bin.' for package '.$package->getName().': file not found in package</warning>');
                 continue;
             }
+
+            // in case a custom installer returned a relative path for the
+            // $package, we can now safely turn it into a absolute path (as we
+            // already checked the binary's existence). The following helpers
+            // will require absolute paths to work properly.
+            $binPath = realpath($binPath);
 
             $this->initializeBinDir();
             $link = $this->binDir.'/'.basename($bin);
@@ -189,19 +214,24 @@ class LibraryInstaller implements InstallerInterface
                     // likely leftover from a previous install, make sure
                     // that the target is still executable in case this
                     // is a fresh install of the vendor.
-                    chmod($link, 0777 & ~umask());
+                    @chmod($link, 0777 & ~umask());
                 }
-                $this->io->write('    Skipped installation of '.$bin.' for package '.$package->getName().': name conflicts with an existing file');
+                $this->io->write('    Skipped installation of bin '.$bin.' for package '.$package->getName().': name conflicts with an existing file');
                 continue;
             }
             if (defined('PHP_WINDOWS_VERSION_BUILD')) {
                 // add unixy support for cygwin and similar environments
                 if ('.bat' !== substr($binPath, -4)) {
                     file_put_contents($link, $this->generateUnixyProxyCode($binPath, $link));
-                    chmod($link, 0777 & ~umask());
+                    @chmod($link, 0777 & ~umask());
                     $link .= '.bat';
+                    if (file_exists($link)) {
+                        $this->io->write('    Skipped installation of bin '.$bin.'.bat proxy for package '.$package->getName().': a .bat proxy was already installed');
+                    }
                 }
-                file_put_contents($link, $this->generateWindowsProxyCode($binPath, $link));
+                if (!file_exists($link)) {
+                    file_put_contents($link, $this->generateWindowsProxyCode($binPath, $link));
+                }
             } else {
                 $cwd = getcwd();
                 try {
@@ -209,13 +239,15 @@ class LibraryInstaller implements InstallerInterface
                     // when using it in smbfs mounted folder
                     $relativeBin = $this->filesystem->findShortestPath($link, $binPath);
                     chdir(dirname($link));
-                    symlink($relativeBin, $link);
+                    if (false === symlink($relativeBin, $link)) {
+                        throw new \ErrorException();
+                    }
                 } catch (\ErrorException $e) {
                     file_put_contents($link, $this->generateUnixyProxyCode($binPath, $link));
                 }
                 chdir($cwd);
             }
-            chmod($link, 0777 & ~umask());
+            @chmod($link, 0777 & ~umask());
         }
     }
 
@@ -227,11 +259,11 @@ class LibraryInstaller implements InstallerInterface
         }
         foreach ($binaries as $bin) {
             $link = $this->binDir.'/'.basename($bin);
-            if (file_exists($link)) {
-                unlink($link);
+            if (is_link($link) || file_exists($link)) {
+                $this->filesystem->unlink($link);
             }
             if (file_exists($link.'.bat')) {
-                unlink($link.'.bat');
+                $this->filesystem->unlink($link.'.bat');
             }
         }
     }
@@ -265,7 +297,7 @@ class LibraryInstaller implements InstallerInterface
         }
 
         return "@ECHO OFF\r\n".
-            "SET BIN_TARGET=%~dp0\\".escapeshellarg(dirname($binPath)).'\\'.basename($binPath)."\r\n".
+            "SET BIN_TARGET=%~dp0/".trim(ProcessExecutor::escape($binPath), '"')."\r\n".
             "{$caller} \"%BIN_TARGET%\" %*\r\n";
     }
 
@@ -276,7 +308,7 @@ class LibraryInstaller implements InstallerInterface
         return "#!/usr/bin/env sh\n".
             'SRC_DIR="`pwd`"'."\n".
             'cd "`dirname "$0"`"'."\n".
-            'cd '.escapeshellarg(dirname($binPath))."\n".
+            'cd '.ProcessExecutor::escape(dirname($binPath))."\n".
             'BIN_TARGET="`pwd`/'.basename($binPath)."\"\n".
             'cd "$SRC_DIR"'."\n".
             '"$BIN_TARGET" "$@"'."\n";
